+++ src/SFA.DAS.EmployerIncentives.ValueObjects/NewApprenticeIncentive.cs
@@ -7,8 +7,8 @@ namespace SFA.DAS.EmployerIncentives.ValueObjects
     public class NewApprenticeIncentive : ValueObject
     {
         private readonly DateTime EligibilityStartDate = new DateTime(2020, 8, 1);
-        private const double TwentyFiveOrOverIncentiveAmount = 2000;
-        private const double UnderTwentyFiveIncentiveAmount = 1500;
+        private const decimal TwentyFiveOrOverIncentiveAmount = 2000;
+        private const decimal UnderTwentyFiveIncentiveAmount = 1500;
 
         public bool IsApprenticeshipEligible(Apprenticeship apprenticeship)
         {
@@ -20,7 +20,7 @@ namespace SFA.DAS.EmployerIncentives.ValueObjects
             return true;
         }
 
-        public double CalculateTotalIncentiveAmount(DateTime apprenticeDateOfBirth, DateTime plannedStartDate)
+        public decimal CalculateTotalIncentiveAmount(DateTime apprenticeDateOfBirth, DateTime plannedStartDate)
         {
             var apprenticeAge = CalculateAgeAtStartOfApprenticeship(apprenticeDateOfBirth, plannedStartDate);
